clean up tech soup
