+++ src/lib/components/TechnologySoup.svelte
@@ -6,15 +6,20 @@
     const techAccumulation: Record<string, number>[] = [];
     let canvas: HTMLCanvasElement;
     // technologies.forEach(t => t in techAccumulation.keys ? techAccumulation += 1 : techAccumulation. )
-
+    let wordSoup: WordSoup;
     onMount(() => {
         canvas.height = 1000;
         canvas.width = canvas.height * (16 / 9);
-        const wordSoup = new WordSoup(canvas, technologies.slice(0, 3));
-        wordSoup.animate();
+        wordSoup = new WordSoup(canvas, technologies);
+        wordSoup.animate(0);
     });
+
+    const toggleRandomForces = () => {
+        wordSoup.randomForces = !wordSoup.randomForces;
+    }
 </script>
 
+<button on:click={toggleRandomForces}>ToggleRandom</button>
 <canvas bind:this={canvas}>
 </canvas>
